Make balances `u128` (#2498)
* WIP

* WIP: It compiles

* Migrate gas price to u128

* Fix priority fee math. Serialize gas price as string

* Update auto-generated files

* Update changelog

* Fix compilation of excluded crates

* WIP: Migrate all balances to Amount struct

* Minor fix in rewards

* Make from_primitive const, add Amount::MAX

* Some progress for macro

* One error!

* Keep going

* Keep spreading amount

* Compiles!

* Keep updating tests

* Making progress

* Making progress part 2

* Lint passes!

* Clean up and satisfy CI

* More tests are passing

* TRy this

* Fix openapi spec

* Updated chain hash

* Insta review

* Fmt and happy switcheroo

* Print key in error!

* More logging

* Fmt and useless conversion

* Remove useless conversion in Genesis

* Ignore max_fee tests

* Lint passes again after merge

* Revert switcheroo

* Use `Amount::MAX`

* Saturating gas_used delta

* Proper feature gating of sov-uniqueness

* Clean up trait

* More clean ups

* Use u64 from gas values

* Try to temp fix genesis configs fro demo data

* After merge

* Chain hash is updated

* Switcherooo, my friend!

* Update with the latest PR update in web3-js

* Fixing demo-rollup readme's, hopefully

* Decompose gas price update into limbs

* Use amount for bank and prover incentives genesis

* Incorporing latest gas charge patterns (#2502)

* Incorporing latest gas charge patterns

* PR comments

* Docs + testfix

* CHANGELOG

* nit

* PR comments

* Fix

* Remove test after bad merge

* Try fixes!

* Fix fmt after merge

* two more fixes for demo-rollup

* Another fix for mock-demo rollup

---------

+++ crates/module-system/module-implementations/sov-paymaster/src/policies.rs
@@ -36,7 +36,7 @@ pub enum PayeePolicy<S: Spec> {
     /// In all other cases, the sender pays their own fees.
     Allow {
         #[allow(missing_docs)]
-        max_fee: Option<u64>,
+        max_fee: Option<u128>,
         #[allow(missing_docs)]
         gas_limit: Option<S::Gas>,
         #[allow(missing_docs)]
@@ -80,7 +80,7 @@ impl<S: Spec> PayeePolicy<S> {
     }
 
     /// Checks that the transaction's max fee is less than the policy's max fee, if applicable.
-    pub fn authorizes_max_fee(&self, tx_max_fee: u64) -> bool {
+    pub fn authorizes_max_fee(&self, tx_max_fee: u128) -> bool {
         // Use `match` instead of `if let` to ensure exhaustive pattern
         match self {
             PayeePolicy::Allow { max_fee, .. } => {
@@ -154,7 +154,7 @@ impl<S: Spec> PayeePolicy<S> {
         }
     }
 
-    fn max_fee(&self) -> Option<u64> {
+    fn max_fee(&self) -> Option<u128> {
         match self {
             PayeePolicy::Allow { max_fee, .. } => *max_fee,
             PayeePolicy::Deny => None,
